Simplify how Seedlet child themes declare color annotations
