+++ spearhead/inc/wpcom-colors-editor.php
@@ -1,0 +1,2 @@
+<?php
+	require_once 'wpcom-colors.php';
